requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-filter
drf-spectacular
psycopg2-binary
stripe